split and fix nano action
