+++ .github/workflows/nano_unit_tests_pytorch.yml
@@ -99,6 +99,22 @@ jobs:
         env:
           ANALYTICS_ZOO_ROOT: ${{ github.workspace }}
 
+  nano-unit-test-pytorch-deps:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ["ubuntu-20.04"]
+        python-version: ["3.7"]
+        pytorch-version: [
+          "pytorch_110 neural-compressor==1.13.1",
+          "pytorch_111 neural-compressor==1.13.1",
+          "pytorch_112 neural-compressor==1.13.1",
+          "pytorch_113 neural-compressor==2.0",
+          ]
+    steps:
+      - uses: actions/checkout@v3
+
       - name: Run PyTorch unit tests(Ray)
         shell: bash
         run: |
