+++ python/nano/test/openvino/tf/test_openvino_quantize.py
@@ -32,7 +32,7 @@ class TestOpenVINO(TestCase):
         train_dataset = tf.data.Dataset.from_tensor_slices((train_examples, train_labels))
 
         # Case1: Trace and quantize
-        openvino_model = model.trace(accelerator='openvino')
+        openvino_model = InferenceOptimizer.trace(model, accelerator='openvino')
         openvino_quantized_model = InferenceOptimizer.quantize(openvino_model,
                                                                accelerator='openvino',
                                                                x=train_dataset,
@@ -52,19 +52,24 @@ class TestOpenVINO(TestCase):
         y_hat = openvino_quantized_model(train_examples[:10])
         assert y_hat.shape == (10, 10)
 
-        y_hat = openvino_quantized_model.predict(train_examples, batch_size=5)
-        assert y_hat.shape == (100, 10)
-
         preds = model.predict(train_examples)
         openvino_preds = openvino_quantized_model.predict(train_examples)
         np.testing.assert_allclose(preds, openvino_preds, rtol=1e-2)
 
         # Case 3: with config
-        openvino_quantized_model = InferenceOptimizer.quantize(model,
+        openvino_quantized_model = InferenceOptimizer.quantize(openvino_model,
                                                                accelerator='openvino',
                                                                x=train_dataset,
                                                                openvino_config={"PERFORMANCE_HINT": "LATENCY"})
+        y_hat = openvino_quantized_model(train_examples[:10])
+        assert y_hat.shape == (10, 10)
 
+        # Case 4: with kargs
+        openvino_quantized_model = InferenceOptimizer.quantize(openvino_model,
+                                                               accelerator='openvino',
+                                                               x=train_dataset,
+                                                               thread_num=8,
+                                                               mean_value=[123.68,116.78,103.94]) # mo param
         y_hat = openvino_quantized_model(train_examples[:10])
         assert y_hat.shape == (10, 10)
 
@@ -150,53 +155,28 @@ class TestOpenVINO(TestCase):
         train_labels = np.random.randint(0, 10, size=(100,))
         train_dataset = tf.data.Dataset.from_tensor_slices((train_examples, train_labels))
 
-        # Case1: Trace and quantize
-        openvino_model = model.trace(accelerator='openvino')
-        openvino_quantized_model = InferenceOptimizer.quantize(openvino_model,
+        openvino_quantized_model = InferenceOptimizer.quantize(model,
                                                                accelerator='openvino',
                                                                x=train_dataset,
                                                                thread_num=8)
-
-        y_hat = openvino_quantized_model(train_examples[:10])
-        assert y_hat.shape == (10, 10)
-
-        y_hat = openvino_quantized_model.predict(train_examples, batch_size=5)
-        assert y_hat.shape == (100, 10)
         
         # test original save / load
         with tempfile.TemporaryDirectory() as tmp_dir_name:
-            openvino_model._save(tmp_dir_name)
+            openvino_quantized_model._save(tmp_dir_name)
             new_ov_model = KerasOpenVINOModel._load(tmp_dir_name)
 
-        preds1 = openvino_model(train_examples).numpy()
+        preds1 = openvino_quantized_model(train_examples).numpy()
         preds2 = new_ov_model(train_examples).numpy()
         np.testing.assert_almost_equal(preds1, preds2, decimal=5)
 
         # test InferencOptimizer save / load
         with tempfile.TemporaryDirectory() as tmp_dir_name:
-            InferenceOptimizer.save(openvino_model, tmp_dir_name)
+            InferenceOptimizer.save(openvino_quantized_model, tmp_dir_name)
             new_ov_model = InferenceOptimizer.load(tmp_dir_name, model)
 
-        preds1 = openvino_model(train_examples).numpy()
+        preds1 = openvino_quantized_model(train_examples).numpy()
         preds2 = new_ov_model(train_examples).numpy()
         np.testing.assert_almost_equal(preds1, preds2, decimal=5)
-
-    def test_model_quantize_openvino_kwargs(self):
-        model = MobileNetV2(weights=None, input_shape=[40, 40, 3], classes=10)
-        model = Model(inputs=model.inputs, outputs=model.outputs)
-        train_examples = np.random.random((100, 40, 40, 3))
-        train_labels = np.random.randint(0, 10, size=(100,))
-        train_dataset = tf.data.Dataset.from_tensor_slices((train_examples, train_labels))
-
-        openvino_model = model.trace(accelerator='openvino')
-        openvino_quantized_model = InferenceOptimizer.quantize(openvino_model,
-                                                               accelerator='openvino',
-                                                               x=train_dataset,
-                                                               thread_num=8,
-                                                               mean_value=[123.68,116.78,103.94]) # mo param
-
-        y_hat = openvino_quantized_model(train_examples[:10])
-        assert y_hat.shape == (10, 10)
     
     def test_model_quantize_openvino_vpu(self):
         # test whether contains VPU
@@ -215,7 +195,7 @@ class TestOpenVINO(TestCase):
         train_labels = np.random.randint(0, 10, size=(100,))
         train_dataset = tf.data.Dataset.from_tensor_slices((train_examples, train_labels))
 
-        openvino_model = model.trace(accelerator='openvino')
+        openvino_model = InferenceOptimizer.trace(model, accelerator='openvino')
         openvino_quantized_model = InferenceOptimizer.quantize(openvino_model,
                                                                accelerator='openvino',
                                                                x=train_dataset,
