+++ python/nano/test/tf/train/test_embeddings.py
